Add option "--driver-mode" to config driver mode used by clang-tidy.
clang-tidy can't specify driver mode from command line. It will cause
problems when it is used with visual studio compiler (cl.exe). For
example, option of cl.exe can be started with either '-' or '/'. But
with default driver mode, you can only use '-' with clang-tidy.

	cl.exe a.cpp /Iinclude

can be success, but

	clang-tidy.exe a.cpp /Iinclude

may fail because /Iinclude can't be recongized by default driver mode of
clang-tidy. You need change the above line to:

	clang-tidy.exe a.cpp -Iinclude

One workaround is passing "--driver-mode=cl" after "--":

	clang-tidy.exe a.cpp -- --driver-mode

But this is not ideal. Generally, the options follow "--" should be
compiler options. "--driver-mode" is not the option of cl.exe. cl.exe
will generate "unknown option" warning.

This change adding an option "--driver-mode" to clang-tidy itself tries
to solve this problem. User can specify "gcc", "g++", "cpp", and "cl"
four modes. This option is optional. If this option is not specified,
the default driver mode will be used.