Add an option "IgnoreStdFunctionArguments" to bugprone-argument-comment.
Standard library function has different prototypes between different
compilers. For example, memset() in gcc, the prototype might be:

	void* memset(void* dest, int ch, size_t count);

So if we run clang-tidy, it might add comment to the argument like this:

	memset(buffer, /*ch=*/0, count);

But in vc, the prototype might be:

	void* memset(void* _Dst, int _Val, size_t _Size);

So if we run clang-tidy, it might add comment to the argument like this:

	memset(buffer, /*_Val*/0, count);

That means the same code might pass build under one compiler but fail
under another compiler.

This change tries to fix this issue. It adds an option
"IgnoreStdFunctionArguments". If it is true, clang-tidy will try its
best to recongize the standard function, and ignore the argument comment
check on such function declared in system library.